feat: 장바구니 쿠폰 적용 API 및 DTO 변경

+++ backend/main-service/src/main/java/kkakka/mainservice/cart/domain/CartItem.java
@@ -75,7 +75,7 @@ public class CartItem {
     }
 
     public Integer getDiscountedPrice(Coupon coupon) {
-        Integer maxDiscount = product.getMaxDiscount(coupon);
-        return product.getPrice() - maxDiscount;
+        int maxDiscount = (int) product.getMaxDiscount(coupon);
+        return product.getPrice() - (Integer) maxDiscount;
     }
 }
